[Feedback] Remove Feedback from the bottom of the layout.
