+++ lib/nexmo_developer/app/presenters/feedback_presenter.rb
@@ -25,19 +25,4 @@ class FeedbackPresenter
   def passed_invisible_captcha?
     @session[:user_passed_invisible_captcha] || false
   end
-
-  def show_github_link?
-    !!@document_path
-  end
-
-  def path_to_url
-    @document_path&.gsub("#{Rails.configuration.docs_base_path}/", '')
-  end
-
-  def github_url
-    @github_url ||= begin
-      info = YAML.safe_load(File.open("#{Rails.configuration.docs_base_path}/config/business_info.yml"))
-      "https://github.com/#{info['docs_repo']}/blob/#{ENV.fetch('branch', 'master')}/#{path_to_url}"
-    end
-  end
 end
