+++ lib/nexmo_developer/app/views/layouts/api.html.erb
@@ -17,7 +17,7 @@
       <%= render partial: 'layouts/partials/notices' %>
       <div class="Vlt-card" id="primary-content">
         <%= yield %>
-        <%= render partial: 'layouts/partials/feedback' %>
+        <%= render partial: 'layouts/partials/improve_this_page' %>
       </div>
     </main>
   </body>
