+++ lib/nexmo_developer/app/views/layouts/documentation.html.erb
@@ -18,7 +18,7 @@
       <div class="Vlt-article">
         <div id="primary-content" tabindex="-1"></div>
         <%= yield %>
-        <%= render partial: 'layouts/partials/feedback' unless @skip_feedback %>
+        <%= render partial: 'layouts/partials/improve_this_page' %>
       </div>
     </main>
   </body>
