+++ lib/nexmo_developer/app/views/layouts/static.html.erb
@@ -26,7 +26,7 @@
       <%= render partial: 'layouts/partials/notices' %>
       <div class="Vlt-card">
         <%= yield %>
-        <%= render partial: 'layouts/partials/feedback' %>
+        <%= render partial: 'layouts/partials/improve_this_page' %>
       </div>
     </main>
   </body>
